requirements.txt
actionlib==1.13.2
alabaster==0.7.8
angles==1.9.13
apturl==0.5.2
attrs==19.3.0
autobahn>=20.12.3
Automat==0.8.0
Babel==2.6.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
bondpy==1.8.6
breezy==3.0.2
Brlapi==0.7.0
camera-calibration==1.15.3
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.4.23
catkin-pkg-modules==0.4.23
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cryptography>=3.2
cupshelpers==1.0
cv-bridge==1.15.0
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
diagnostic-analysis==1.10.4
diagnostic-common-diagnostics==1.10.4
diagnostic-updater==1.10.4
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
dulwich==0.19.15
dynamic-reconfigure==1.7.1
empy==3.3.2
entrypoints==0.3
fastimport==0.9.8
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
gpg===1.13.1-unknown
html5lib==1.0.1
httplib2==0.19.0
hyperlink==19.0.0
idna==2.8
image-geometry==1.15.0
imagesize==1.2.0
incremental==16.10.1
interactive-markers==1.12.0
Jinja2==2.10.1
joint-state-publisher==1.15.0
joint-state-publisher-gui==1.15.0
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml>=4.6.3
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
MarkupSafe==1.1.0
matplotlib==3.1.2
message-filters==1.15.11
moveit-commander==1.1.5
moveit-core==1.1.5
moveit-ros-planning-interface==1.1.5
moveit-ros-visualization==1.1.5
mpi4py==3.0.3
netifaces==0.10.4
nose==1.3.7
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
paramiko==2.6.0
pexpect==4.6.0
Pillow>=8.2.0
protobuf==3.6.1
psutil>=5.6.6
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyassimp==4.1.4
pycairo==1.16.2
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
PyGithub==1.43.7
Pygments>=2.7.4
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyrealsense2==2.48.0.3381
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-qt-binding==0.4.4
python-snappy==0.5.3
PyTrie==0.2
pyttsx3==2.90
pytz==2019.3
pyxdg==0.26
PyYAML>=5.4
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
resource-retriever==1.12.6
roman==2.0.0
rosbag==1.15.11
rosboost-cfg==1.15.7
rosclean==1.15.7
roscreate==1.15.7
rosdep==0.21.0
rosdep-modules==0.21.0
rosdistro==0.8.3
rosdistro-modules==0.8.3
rosgraph==1.15.11
rosinstall==0.7.8
rosinstall-generator==0.1.22
roslaunch==1.15.11
roslib==1.15.7
roslint==0.12.0
roslz4==1.15.11
rosmake==1.15.7
rosmaster==1.15.11
rosmsg==1.15.11
rosnode==1.15.11
rosparam==1.15.11
rospkg==1.3.0
rospkg-modules==1.3.0
rospy==1.15.11
rosservice==1.15.11
rostest==1.15.11
rostopic==1.15.11
rosunit==1.15.7
roswtf==1.15.11
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.2
rqt-py-console==0.4.10
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.2
rqt-top==0.4.10
rqt-topic==0.4.12
rqt-web==0.4.10
rviz==1.14.8
scipy==1.7.1
SecretStorage==2.3.1
sensor-msgs==1.13.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
soupsieve==1.9.5
SpeechRecognition==3.8.1
Sphinx==1.8.5
sphinx-rtd-theme==0.4.3
srdfdom==0.6.2
systemd-python==234
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
topic-tools==1.15.11
transforms3d==0.3.1
Twisted==18.9.0
txaio==2.10.0
u-msgpack-python==2.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urdfdom-py==0.4.5
urllib3>=1.26.5
vcstools==0.1.42
wadllib==1.3.3
webencodings==0.5.1
wrapt==1.11.2
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.8
xkit==0.0.0
zope.interface==4.7.1
getch==1.0